Work around compiler bug
As detailed at http://bugs.sun.com/view_bug.do?bug_id=6682380, using
a generic wildcard in a foreach loop within a finally block causes
a compiler error.  This has been fixed in Sun's JDK, but not apparently
by Apple. Confirmed the error still occurs on javac 1.6.0_20 and _22.

The workaround is to simply remove the generic wildcard (from the
Class loop variable in this case), and add @SuppressWarnings("rawtypes")
to avoid the ensuing compiler warning.